requirements.txt
aiofiles
amqp==5.3.1
async-timeout==5.0.1
bidict==0.23.1
billiard==4.2.1
blinker
Brotli
celery==5.4.0
certifi
charset-normalizer
click
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
eventlet==0.38.0
Flask
Flask-Compress==1.17
Flask-Cors
Flask-SocketIO==5.4.1
gevent==24.11.1
gevent-websocket==0.10.1
greenlet
h11==0.14.0
idna
importlib_metadata
itsdangerous
Jinja2
kombu==5.4.2
MarkupSafe
mysql-connector-python==9.1.0
numpy==2.0.2
opencv-python==4.10.0.84
pillow==11.0.0
prompt_toolkit==3.0.48
PyMySQL==1.1.1
PySocks
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
redis==5.2.0
requests
simple-websocket==1.1.0
six
SQLAlchemy
tqdm==4.67.1
typing_extensions
tzdata==2024.2
urllib3
vine==5.1.0
wcwidth==0.2.13
Werkzeug
wsproto==1.2.0
zipp
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
